Refactor Formula Form component to make it reusable outside Formulas UI
