+++ web/html/src/components/formulas/formula-form.css
@@ -52,7 +52,7 @@
     color: #8e8e90;
 }
 
-:global(.formula-content :invalid) {
+:global(input:invalid) {
     background-color: #fcf8e3;
 }
